requirements.txt
python-dotenv
pandas
genanki